+++ tools/installPSResources.ps1
@@ -5,6 +5,7 @@ param(
     [string]$PSRepository = "PSGallery"
 )
 
+# Install-PSResource can't use the project-scoped feed because OneBranch doesn't auth it
 if ($PSRepository -eq "CFS" -and -not (Get-PSResourceRepository -Name CFS -ErrorAction SilentlyContinue)) {
     Register-PSResourceRepository -Name CFS -Uri "https://pkgs.dev.azure.com/powershell/PowerShell/_packaging/PowerShellGalleryMirror/nuget/v3/index.json"
 }
@@ -12,11 +13,11 @@ if ($PSRepository -eq "CFS" -and -not (Get-PSResourceRepository -Name CFS -Error
 # NOTE: Due to a bug in Install-PSResource with upstream feeds, we have to
 # request an exact version. Otherwise, if a newer version is available in the
 # upstream feed, it will fail to install any version at all.
-Install-PSResource -Verbose -TrustRepository -RequiredResource  @{
+Install-PSResource -Verbose -TrustRepository -RequiredResource @{
     InvokeBuild = @{
-        version = "5.12.1"
+        version = "5.14.23"
         repository = $PSRepository
-      }
+    }
     platyPS = @{
         version = "0.14.2"
         repository = $PSRepository